requirements.txt
jax-ai-stack[tfds]
tensorflow-cpu
tensorflow-datasets>=4.9.0
orbax-checkpoint>=0.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
numpy>=1.24.0
